Grid, voltage scale, and time scale are drawn
Grid, voltage scale, and time scale are drawn but signal is still not showing. Part of problem is the code seems to be jumping between processing and display task but never reaches the waveform task. To test this, place breakpoints at the beginning of each function and run the code.